+++ Evergreen/Private/Invoke-WebContent.ps1
@@ -23,7 +23,10 @@ Function Invoke-WebContent {
         [System.String] $UserAgent = [Microsoft.PowerShell.Commands.PSUserAgent]::Chrome,
 
         [Parameter()]
-        [System.Management.Automation.SwitchParameter] $Raw
+        [System.Management.Automation.SwitchParameter] $Raw,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter] $TrustCertificate
     )
 
     # Use TLS 1.2
@@ -44,6 +47,26 @@ Function Invoke-WebContent {
     }
     Else {
         $errorAction = "SilentlyContinue"
+    }
+
+    If ($TrustCertificate.IsPresent) {
+        If (Test-PSCore) {
+            Write-Warning -Message "$($MyInvocation.MyCommand): Running PowerShell Core. Skipping System.Security.Cryptography.X509Certificates."
+        }
+        Else {
+            Add-Type @"
+using System.Net;
+using System.Security.Cryptography.X509Certificates;
+public class TrustAllCertsPolicy : ICertificatePolicy {
+    public bool CheckValidationResult(
+        ServicePoint srvPoint, X509Certificate certificate,
+        WebRequest request, int certificateProblem) {
+        return true;
+    }
+}
+"@
+            [System.Net.ServicePointManager]::CertificatePolicy = New-Object TrustAllCertsPolicy
+        }
     }
 
     try {