+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## VERSION
+
+* Adds back `Get-FileZilla` using the application update API. Currently returns only the 64-bit version of FileZilla for Windows.
+
 ## 2004.125
 
 * Adds `Get-MicrosoftOneDrive`. We recommend validating versions returned by this function with [OneDrive release notes](https://support.office.com/en-us/article/onedrive-release-notes-845dcf18-f921-435e-bf28-4e24b95e5fc0)